requirements.txt
attrs==16.3.0
mock==2.0.0
requests==2.11.1
beautifulsoup4==4.5.1
